+++ changelog.md
@@ -1,3 +1,9 @@
+== NoMoreRecipeConflict 0.2 ==
+* Added cycle button in crafting gui (enabled by default)
+* Changed switch key to disabled by default
+* Added config options to choose between key or button
+* Added extra checks server side to avoid trouble
+
 == NoMoreRecipeConflict 0.1 ==
 * Port to Forge
 * Removed all base class edits